test(reborn): extract mock-MCP scaffolding into harness_mcp.rs (W0-SPLIT)
Behavior-preserving structural refactor. Moves the cohesive mock-MCP block
out of the 4.4k-line tests/support/reborn/harness.rs into a focused
harness_mcp.rs submodule so later coverage enablers target focused files:

- LoopbackMcpRuntimeHttpEgress (real-HTTP loopback egress) + impls
- LoopbackMcpRuntime alias + build_loopback_mcp_runtime factory
- mock_mcp_extension_package, local_dev_host_runtime_with_registry_egress_and_mcp
- MCP trust + loopback network policies

HostRuntimeCapabilityHarness::mock_mcp_tools stays in harness.rs (a full
Self{} constructor co-located with its sibling constructors; moving it would
force all 23 private fields of the central harness struct to widen). It now
delegates the MCP wiring to the pub(super) factories in harness_mcp.rs, so
harness.rs no longer imports ironclaw_mcp at all.

Widens three shared harness internals to pub(crate) (RecordingRuntimeHttpEgress,
local_dev_root_filesystem, LocalDevRootMounts) for the new sibling module.
No test logic changes; reborn_integration_mcp test set unchanged (19 tests).

harness.rs 4387 -> 4076 lines; harness_mcp.rs 367 lines.
